requirements.txt
selenium==4.11.2
PyYAML==6.0.1
mitmproxy==10.0.0
bs4==0.0.1
beautifulsoup4==4.12.2